refactor: move element source span extension.
