+++ tools/serverpod_cli/lib/src/analyzer/dart/element_extensions.dart
@@ -1,0 +1,21 @@
+import 'package:analyzer/dart/element/element.dart';
+import 'package:source_span/source_span.dart';
+
+extension DartElementSourceSpan on Element {
+  SourceSpan? get span {
+    var sourceData = source?.contents.data;
+    var sourceUri = source?.uri;
+    var offset = nameOffset;
+    var length = nameLength;
+
+    if (sourceData != null && offset != 0 && length != -1) {
+      var sourceFile = SourceFile.fromString(
+        sourceData,
+        url: sourceUri,
+      );
+      return sourceFile.span(offset, offset + length);
+    } else {
+      return null;
+    }
+  }
+}
